Fixed linting with SwiftLint, Fixed typos and formatting with SwiftFormat. Fixed cocoapods to include Receipt in Example
