+++ Sources/Printer/Hardware/Bluetooth/BluetoothCentralManagerDelegate.swift
@@ -6,27 +6,26 @@
 //  Copyright © 2016 Kevin. All rights reserved.
 //
 
-import Foundation
 import CoreBluetooth
+import Foundation
 
 class BluetoothCentralManagerDelegate: NSObject, CBCentralManagerDelegate {
-
-    struct UserDefaultKey {
-
-        static let autoConectUUID = "auto.connect.uuid"
+    enum UserDefaultKey {
+        static let autoConnectUUID = "auto.connect.uuid"
     }
 
     private var services: Set<String>!
 
     var peripheralDelegate: BluetoothPeripheralDelegate?
 
-    open var centralManagerDidUpdateState: ((CBCentralManager) -> ())?
-    open var centralManagerDidDiscoverPeripheralWithAdvertisementDataAndRSSI: ((CBCentralManager, CBPeripheral, [String : Any], NSNumber) -> ())?
-    open var centralManagerDidConnectPeripheral: ((CBCentralManager, CBPeripheral) -> ())?
-    open var centralManagerDidFailToConnectPeripheralWithError: ((CBCentralManager, CBPeripheral, Error?) -> ())?
-    open var centralManagerDidDisConnectPeripheralWithError: ((CBCentralManager, CBPeripheral, Error?) -> ())?
+    open var centralManagerDidUpdateState: ((CBCentralManager) -> Void)?
+    open var centralManagerDidDiscoverPeripheralWithAdvertisementDataAndRSSI:
+        ((CBCentralManager, CBPeripheral, [String: Any], NSNumber) -> Void)?
+    open var centralManagerDidConnectPeripheral: ((CBCentralManager, CBPeripheral) -> Void)?
+    open var centralManagerDidFailToConnectPeripheralWithError: ((CBCentralManager, CBPeripheral, Error?) -> Void)?
+    open var centralManagerDidDisConnectPeripheralWithError: ((CBCentralManager, CBPeripheral, Error?) -> Void)?
 
-    typealias PeripheralChangeBlock = (UUID) -> ()
+    typealias PeripheralChangeBlock = (UUID) -> Void
 
     var addedPeripherals: PeripheralChangeBlock?
     var updatedPeripherals: PeripheralChangeBlock?
@@ -41,7 +40,6 @@ class BluetoothCentralManagerDelegate: NSObject, CBCentralManagerDelegate {
             return discoveredPeripherals[uuid]
         }
         set {
-
             let oldValue = discoveredPeripherals[uuid]?.identifier
 
             lock.lock()
@@ -68,7 +66,6 @@ class BluetoothCentralManagerDelegate: NSObject, CBCentralManagerDelegate {
     }
 
     public func centralManagerDidUpdateState(_ central: CBCentralManager) {
-
         centralManagerDidUpdateState?(central)
 
         let ss = services.map { CBUUID(string: $0) }
@@ -80,25 +77,28 @@ class BluetoothCentralManagerDelegate: NSObject, CBCentralManagerDelegate {
         }
     }
 
-    public func centralManager(_ central: CBCentralManager, didDiscover peripheral: CBPeripheral, advertisementData: [String : Any], rssi RSSI: NSNumber) {
-
+    public func centralManager(_ central: CBCentralManager,
+                               didDiscover peripheral: CBPeripheral,
+                               advertisementData: [String: Any],
+                               rssi RSSI: NSNumber)
+    {
         guard let serviceUUIDs = advertisementData[CBAdvertisementDataServiceUUIDsKey] as? [CBUUID],
-            let isConnectable = advertisementData[CBAdvertisementDataIsConnectable] as? NSNumber,
-            serviceUUIDs.count > 0, isConnectable == 1 else {
-                return
+              let isConnectable = advertisementData[CBAdvertisementDataIsConnectable] as? NSNumber,
+              serviceUUIDs.count > 0, isConnectable == 1
+        else {
+            return
         }
 
         // if peripheral doesn't container specified services, ignore.
-        let peripheralServiceSet = Set(serviceUUIDs.map { $0.uuidString } )
+        let peripheralServiceSet = Set(serviceUUIDs.map { $0.uuidString })
 
         guard peripheralServiceSet.intersection(services).count > 0 else {
-
             return
         }
 
         self[peripheral.identifier] = peripheral
 
-        if let uuid = UserDefaults.standard.object(forKey: UserDefaultKey.autoConectUUID) as? String {
+        if let uuid = UserDefaults.standard.object(forKey: UserDefaultKey.autoConnectUUID) as? String {
             if peripheral.identifier.uuidString == uuid {
                 central.connect(peripheral, options: nil)
             }
@@ -107,25 +107,27 @@ class BluetoothCentralManagerDelegate: NSObject, CBCentralManagerDelegate {
         centralManagerDidDiscoverPeripheralWithAdvertisementDataAndRSSI?(central, peripheral, advertisementData, RSSI)
     }
 
-    public func centralManager(_ central: CBCentralManager, didConnect peripheral: CBPeripheral) {
-
+    public func centralManager(_ central: CBCentralManager,
+                               didConnect peripheral: CBPeripheral)
+    {
         centralManagerDidConnectPeripheral?(central, peripheral)
 
         peripheral.delegate = peripheralDelegate
         peripheral.discoverServices(services.map { CBUUID(string: $0) })
 
         // 保存已连接的uuid
-        UserDefaults.standard.set(peripheral.identifier.uuidString, forKey: UserDefaultKey.autoConectUUID)
+        UserDefaults.standard.set(peripheral.identifier.uuidString, forKey: UserDefaultKey.autoConnectUUID)
         UserDefaults.standard.synchronize()
     }
 
     public func centralManager(_ central: CBCentralManager, didFailToConnect peripheral: CBPeripheral, error: Error?) {
-
         centralManagerDidFailToConnectPeripheralWithError?(central, peripheral, error)
     }
 
-    public func centralManager(_ central: CBCentralManager, didDisconnectPeripheral peripheral: CBPeripheral, error: Error?) {
-
+    public func centralManager(_ central: CBCentralManager,
+                               didDisconnectPeripheral peripheral: CBPeripheral,
+                               error: Error?)
+    {
         centralManagerDidDisConnectPeripheralWithError?(central, peripheral, error)
     }
 }
